requirements.txt
pyautogui
pyperclip
pyscreeze
Pillow
PyYAML